requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
django==4.2.25
django-filter==24.3
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.6
graphql-relay==3.2.0
promise==2.3
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
typing-extensions==4.13.2
django-crontab

